+++ changelog.md
@@ -6,7 +6,7 @@
 
 - Updated Node.js support to `^12.20 || >= 14.13`.
 - Updated dependencies, some of which require newer Node.js versions than were previously supported.
-- The tests are now ESM in an `.mjs` file instead of CJS in a `.js` file.
+- The tests are now ESM in `.mjs` files instead of CJS in a `.js` file.
 
 ### Patch